AoC16 - migration from Java 11 to Java 17 LTS (14.5 seconds to 14.2 seconds)

+++ AoC16/src/test/java/aoc16/d12/Day12Test.java
@@ -73,6 +73,6 @@ public class Day12Test {
             final var ci = code.get(i);
             i += ci.getValue1().apply( ci.getValue0(), code, i, registers );
         }
-        return registers.get('a').intValue();
+        return registers.get('a');
     }
 }
